refactor(controller): fix schema imports
Refs #126
