+++ backend/src/main/controllers/transaction/updateTransactionController.ts
@@ -13,7 +13,7 @@ import { formatISO, parseISO } from "date-fns";
 import { IUpdateTransactionUseCase } from "@/main/config/helpers/useCases/IuseCases";
 import { TransactionType } from "@/domain/entities/Transaction";
 import { UpdateTransactionResponse } from "@/main/config/helpers/protocol/transaction/updateTransactionProtocols";
-import { UpdateTransactionSchema } from "@/application/services/updateTransactionSchema";
+import { UpdateTransactionSchema } from "@/application/schema/transaction/updateTransactionSchema";
 import { UpdateTransactionWithCategoryNameParams } from "@/application/interfaces/domain/entities/transaction/ItransactionRepository";
 import { log } from "@/main/config/logs/log";
